--- booknetwork/__init__.py
@@ -1,34 +1,0 @@
-import os
-
-from flask import Flask, session
-from flask_session import Session
-from sqlalchemy import create_engine
-from sqlalchemy.orm import scoped_session, sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-from flask_bootstrap import Bootstrap
-from flask_login import LoginManager
-
-app = Flask(__name__)
-
-# Check for environment variable
-if not os.getenv("DATABASE_URL"):
-    raise RuntimeError("DATABASE_URL is not set")
-
-app.config['SECRET_KEY'] = 'gs3D#00kR3@*^'
-# Configure session to use filesystem
-app.config["SESSION_PERMANENT"] = False
-app.config["SESSION_TYPE"] = "filesystem"
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-Session(app)
-
-#control userlogin
-login_manager = LoginManager()
-login_manager.init_app(app)
-login_manager.login_view = 'login'
-#wtform formatting
-bootstrap = Bootstrap(app)
-# Set up database
-engine = create_engine(os.getenv("DATABASE_URL"))
-db = scoped_session(sessionmaker(bind=engine))
-
-from booknetwork import routes
